requirements.txt
fastapi
uvicorn
python-multipart
starlette
tensorflow
keras
tf_keras
numpy
h5py
pillow
firebase-admin
google-auth
python-dotenv
watchfiles
requests